creating separate teachers firebase file
