+++ src/firebase/teachers.js
@@ -1,0 +1,34 @@
+import { db, checkUserClaim } from './firebase';
+
+export async function updateGroupTeacher(teacherEmail, groupName) {
+  const userIsModerator = await checkUserClaim('moderator');
+  if (!userIsModerator) {
+    return new Error('User not authorized.');
+  }
+  const group = await getGroup(groupName);
+  const groupRef = await db.collection("groups").doc(group);
+  return groupRef.update({
+    teachers: db.FieldValue.arrayUnion(teacherEmail)
+  });
+}
+
+export async function removeGroupTeacher(teacherEmail, groupName) {
+  const userIsModerator = await checkUserClaim('moderator');
+  if (!userIsModerator) {
+    return new Error('User not authorized.');
+  }
+  const group = await getGroup(groupName);
+  const groupRef = await db.collection("groups").doc(group);
+  return groupRef.update({
+    teachers: db.FieldValue.arrayRemove(teacherEmail)
+  });
+}
+
+async function getGroup(groupName) {
+  const querySnapshot = await db.collection("groups").where("groupName", "==", groupName).get();
+  let groups = [];
+  querySnapshot.forEach((doc) => {
+    groups.push(doc.id);
+  });
+  return groups[0];
+}
